backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose[cryptography]==3.5.0
feedparser==6.0.11
aiosmtplib==4.0.2
apscheduler==3.11.0
python-multipart==0.0.20
